Change default container and network prefix names for Koha
Changes in koha-testing-docker have lead to the default naming
convention changing from 'koha' to 'kohadev'. This leads to adb
being unable to connect to the ktd containers without running
ktd with parameters to compensate ( `ktd -—proxy -—name koha up` ).

It would be nice if adb could continue to connect to Koha using
the "out of the box" settings for ktd ( `ktd up` ).

Test Plan:
1) Check out the latest versions of adb and ktd
2) launch both with the default settings ( `ktd up`, `adb up` )
3) Attempt to import Koha data
   `cd /usr/local/aspen-discovery/code/koha_export; java -jar koha_export.jar test.localhostaspen`
4) Note the errors
5) Check out this branch of adb
6) Take down adb and bring it back up ( `adb down`, `adb up` )
7) ADB can talk to Koha again!

Resolves Aspen-Discovery/aspen-dev-box-image#1